+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 - Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
 
-## [1.6.2-beta] - 2026-06-15
+## [1.6.2] - 2026-06-17
 ### Fixed
 - Fix Better Level Maintainer being mistaken for a RandomComplement missing-craft request when `enableMissCraft` is enabled, which could force impossible crafts instead of reporting the failure normally.
 - Fix Pattern/Inventory item rendering eating on the AutoCrafter overview's item renders when the overview is open, by short-circuiting the recipe/inventory rendering when it is open.
@@ -16,6 +16,8 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 
 ### Added
 - Add textures for the Storage Level Emitter and Storage Display.
+- Add subnet support to the Network Health Scanner.
+- Add Subnet Proxy support for anything with subnet support (NACL, NHS).
 
 
 ## [1.6.1-alpha2] - 2026-06-02